Rename program package to check
